+++ NEWS.md
@@ -1,5 +1,7 @@
 # forcats 0.1.1.9000
 
+* `fct_recode()` is now checks that each new value is of length 1 (#56).
+
 # forcats 0.1.1
 
 * Minor fixes for R CMD check